+++ matminer/descriptors/composition_features.py
@@ -1,20 +1,22 @@
+from pymatgen import Element, Composition, MPRester
 import collections
+import os
 import json
 import itertools
 
 import numpy as np
+import pandas as pd
 
-import os
-import re
-from functools import reduce
-
-from pymatgen import Element, Composition, MPRester
-from pymatgen.core.units import Unit
-from pymatgen.core.periodic_table import get_el_sp
+from base_classes import BaseFeaturizer
 
 __author__ = 'Saurabh Bajaj <sbajaj@lbl.gov>'
 
 # TODO: read Magpie file only once
+# TODO: Handle dictionaries in case of atomic radii. Aj says "You can require that getting the ionic_radii descriptor
+#  requires a valence-decorated Structure or valence-decorated Composition. Otherwise it does not work, i.e. returns
+# None. Other radii (e.g. covalent) won't require an oxidation state and people can and should use those for
+# non-ionic structures. You can also have a function that returns a mean of ionic_radii for all valences but that
+# should not be the default."
 # TODO: unit tests
 # TODO: most of this code needs to be rewritten ... AJ
 
@@ -31,68 +33,302 @@ atomic_syms = []
 for atomic_no in range(1,104):
     atomic_syms.append(Element.from_Z(atomic_no).symbol)
 
-def get_pymatgen_descriptor(composition, property_name):
+class MagpieFeaturizer(BaseFeaturizer):
+    """Class to get data from Magpie files"""    
+
+    def __init__(self):
+        BaseFeaturizer.__init__(self)
+
+    def get_data(self, comp_obj, descriptor_name):
+        """
+        Gets magpie data for a composition object. 
+        First checks if magpie properties are already loaded, if not, stores magpie data in a dictionary
+
+        Args: 
+            comp_obj: Pymatgen composition object
+            descriptor_name (string): Name of descriptor
+
+        Returns:
+            magpiedata (list): list of values for each atom in comp_obj
+        """
+
+        if descriptor_name not in magpie_props:
+
+            data_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), "data", 'magpie_elementdata')
+            available_props = []
+
+            # Make a list of available properties
+            for datafile in os.listdir(data_dir):
+                available_props.append(datafile.replace('.table', ''))
+
+            if descriptor_name not in available_props:
+                raise ValueError(
+                    "This descriptor is not available from the Magpie repository. Choose from {}".format(available_props))
+
+            prop_value = []
+            with open(os.path.join(data_dir, '{}.table'.format(descriptor_name)), 'r') as descp_file:
+                lines = descp_file.readlines()
+                
+                for atomic_no in range(1, 104): #This is as high as pymatgen goes
+                    try:
+                        if descriptor_name in ["OxidationStates"]:
+                            prop_value.append([float(i) for i in lines[atomic_no - 1].split()])
+                        else: 
+                            prop_value.append(float(lines[atomic_no - 1]))
+                    except:
+                        prop_value.append(float("NaN"))
+            
+            attr_dict = dict(zip(atomic_syms, prop_value))    
+            magpie_props[descriptor_name] = attr_dict #Add dictionary to magpie_props
+
+        #Get data for given element/compound
+        el_amt = comp_obj.get_el_amt_dict()
+        elements = list(el_amt.keys())
+        
+        magpiedata = []
+        for el in elements:
+            for i in range(int(el_amt[el])):
+                magpiedata.append(magpie_props[descriptor_name][el])
+     
+        return magpiedata
+
+class StoichAttributes(MagpieFeaturizer):
+    """
+    Class to calculate stoichiometric attributes.
+
+    Parameters:
+        p_list (list of ints): list of norms to calculate
+    """
+
+    def __init__(self, p_list=None):
+        MagpieFeaturizer.__init__(self)
+        if p_list == None:
+            self.p_list = [0,2,3,5,7,10]
+        else:
+            self.p_list = p_list
+ 
+    def featurize(self, comp_obj):
+        """
+        Get stoichiometric attributes
+        Args:
+            comp_obj: Pymatgen composition object
+            p_list (list of ints)
+        
+        Returns: 
+            p_norm (float): Lp norm-based stoichiometric attribute
+        """
+ 
+        el_amt = comp_obj.get_el_amt_dict()
+        
+        p_norms = [0]*len(self.p_list)
+        n_atoms = sum(el_amt.values())
+
+        for i in range(len(self.p_list)):
+            if self.p_list[i] < 0:
+                raise ValueError("p-norm not defined for p < 0")
+            if self.p_list[i] == 0:
+                p_norms[i] = len(el_amt.values())
+            else:
+                for j in el_amt:
+                    p_norms[i] += (el_amt[j]/n_atoms)**self.p_list[i]
+                p_norms[i] = p_norms[i]**(1.0/self.p_list[i])
+
+        return p_norms
+
+    def generate_labels(self):
+        labels = []
+        for p in self.p_list:
+            labels.append("%d-norm"%p)
+        return labels
+
+class ElemPropertyAttributes(MagpieFeaturizer):
+    """
+    Class to calculate elemental property attributes
+
+    Parameters:
+        attributes (list of strings): List of elemental properties to use    
+    """
+
+    def __init__(self, attributes=None):
+        MagpieFeaturizer.__init__(self)
+        if attributes == None:
+            self.attributes = ["Number", "MendeleevNumber", "AtomicWeight","MeltingT","Column","Row","CovalentRadius","Electronegativity",
+                "NsValence","NpValence","NdValence","NfValence","NValance","NsUnfilled","NpUnfilled","NdUnfilled","NfUnfilled","NUnfilled",
+                "GSvolume_pa","GSbandgap","GSmagmom","SpaceGroupNumber"]    
+        else: 
+            self.attributes = attributes
+
+    def featurize(self, comp_obj):
+        """
+        Get elemental property attributes
+
+        Args:
+            comp_obj: Pymatgen composition object
+        
+        Returns:
+            all_attributes: min, max, range, mean, average deviation, and mode of descriptors
+        """
+
+        el_amt = comp_obj.get_el_amt_dict()
+        elements = list(el_amt.keys())
+       
+        all_attributes = []
+
+        for attr in self.attributes:
+            
+            elem_data = self.get_data(comp_obj, attr)
+
+            all_attributes.append(min(elem_data))
+            all_attributes.append(max(elem_data))
+            all_attributes.append(max(elem_data) - min(elem_data))
+            
+            prop_mean = sum(elem_data)/len(elem_data)
+            all_attributes.append(prop_mean)
+            all_attributes.append(sum(np.abs(np.subtract(elem_data, prop_mean)))/len(elem_data))
+            all_attributes.append(max(set(elem_data), key=elem_data.count))
+
+        return all_attributes
+
+    def generate_labels(self):
+        labels = []
+        for attr in self.attributes:
+            labels.append("Min %s"%attr)
+            labels.append("Max %s"%attr)
+            labels.append("Range %s"%attr)
+            labels.append("Mean %s"%attr)
+            labels.append("AbsDev %s"%attr)
+            labels.append("Mode %s"%attr)
+        return labels
+
+class ValenceOrbitalAttributes(MagpieFeaturizer):
+    """Class to calculate valence orbital attributes"""
+
+    def __init__(self):
+        MagpieFeaturizer.__init__(self)
+
+    def featurize(self, comp_obj):
+        """Weighted fraction of valence electrons in each orbital
+
+           Args: 
+                comp_obj: Pymatgen composition object
+
+           Returns: 
+                Fs, Fp, Fd, Ff (float): Fraction of valence electrons in s, p, d, and f orbitals
+        """    
+        
+        num_atoms = comp_obj.num_atoms
+
+        avg_total_valence = sum(self.get_data(comp_obj,"NValance"))/num_atoms
+        avg_s = sum(self.get_data(comp_obj,"NsValence"))/num_atoms
+        avg_p = sum(self.get_data(comp_obj,"NpValence"))/num_atoms
+        avg_d = sum(self.get_data(comp_obj,"NdValence"))/num_atoms
+        avg_f = sum(self.get_data(comp_obj,"NfValence"))/num_atoms
+
+        Fs = avg_s/avg_total_valence
+        Fp = avg_p/avg_total_valence
+        Fd = avg_d/avg_total_valence
+        Ff = avg_f/avg_total_valence
+
+        return list((Fs, Fp, Fd, Ff))
+
+    def generate_labels(self):
+        orbitals = ["s","p","d","f"]
+        labels = []
+        for orb in orbitals:
+            labels.append("Frac %s Valence Electrons"%orb)
+
+        return labels
+
+class IonicAttributes(MagpieFeaturizer):
+    """Class to calculate ionic property attributes"""
+    def __init__(self):
+        MagpieFeaturizer.__init__(self)
+
+    def featurize(self, comp_obj):
+        """
+        Ionic character attributes
+
+        Args:
+            com_obj: Pymatgen composition object
+
+        Returns:
+            cpd_possible (bool): Indicates if a neutral ionic compound is possible
+            max_ionic_char (float): Maximum ionic character between two atoms
+            avg_ionic_char (float): Average ionic character
+        """
+
+        el_amt = comp_obj.get_el_amt_dict()
+        elements = list(el_amt.keys())
+        values = list(el_amt.values())
+
+        if len(elements) < 2: #Single element
+            cpd_possible = True
+            max_ionic_char = 0
+            avg_ionic_char = 0        
+        else:
+            #Get magpie data for each element
+            all_ox_states = self.get_data(comp_obj,"OxidationStates")
+            all_elec = self.get_data(comp_obj,"Electronegativity")
+            ox_states = []
+            elec = []
+            
+            for i in range(1,len(values)+1):
+                ind = int(sum(values[:i])-1)
+                ox_states.append(all_ox_states[ind])
+                elec.append(all_elec[ind])
+
+            #Determine if neutral compound is possible
+            cpd_possible = False
+            ox_sets = itertools.product(*ox_states)
+            for ox in ox_sets:
+                if np.dot(ox, values) == 0:
+                    cpd_possible = True
+                    break    
+
+            #Ionic character attributes
+            atom_pairs = itertools.combinations(range(len(elements)), 2)
+            el_frac = list(np.divide(values, sum(values)))
+
+            ionic_char = []
+            avg_ionic_char = 0
+
+            for pair in atom_pairs:
+                XA = elec[pair[0]]
+                XB = elec[pair[1]]
+                ionic_char.append(1.0 - np.exp(-0.25*(XA-XB)**2))
+                avg_ionic_char += el_frac[pair[0]]*el_frac[pair[1]]*ionic_char[-1]
+            
+            max_ionic_char = np.max(ionic_char)
+         
+        return list((cpd_possible, max_ionic_char, avg_ionic_char))
+
+    def generate_labels(self):
+        labels = ["compound possible", "Max Ionic Char", "Avg Ionic Char"]
+        return labels
+
+def get_pymatgen_descriptor(comp, prop):
     """
     Get descriptor data for elements in a compound from pymatgen.
 
     Args:
-        composition (str/Composition): Either pymatgen Composition object or string formula,
-            eg: "NaCl", "Na+1Cl-1", "Fe2+3O3-2" or "Fe2 +3 O3 -2"
-            Notes:
-                 - For 'ionic_radii' property, the Composition object must be made of oxidation
-                    state decorated Specie objects not the plain Element objects.
-                    eg.  fe2o3 = Composition({Specie("Fe", 3): 2, Specie("O", -2): 3})
-                 - For string formula, the oxidation state sign(+ or -) must be specified explicitly.
-                    eg.  "Fe2+3O3-2"
-
-        property_name (str): pymatgen element attribute name, as defined in the Element class at
+        comp: (str) compound composition, eg: "NaCl"
+        prop: (str) pymatgen element attribute, as defined in the Element class at
             http://pymatgen.org/_modules/pymatgen/core/periodic_table.html
 
-    Returns:
-        (list) of values containing descriptor floats for each atom in the compound(sorted by the
-            electronegativity of the contituent atoms)
+    Returns: (list) of values containing descriptor floats for each atom in the compound
 
     """
     eldata = []
-    # what are these named tuples for? not used or returned! -KM
     eldata_tup_lst = []
     eldata_tup = collections.namedtuple('eldata_tup', 'element propname propvalue propunit amt')
+    el_amt_dict = Composition(comp).get_el_amt_dict()
 
-    oxidation_states = {}
-    if isinstance(composition, Composition):
-        # check whether the composition is composed of oxidation state decorates species (not just plain Elements)
-        if hasattr(composition.elements[0], "oxi_state"):
-            oxidation_states = dict([(str(sp.element), sp.oxi_state) for sp in composition.elements])
-        el_amt_dict = composition.get_el_amt_dict()
-    # string
-    else:
-        comp, oxidation_states = get_composition_oxidation_state(composition)
-        el_amt_dict = comp.get_el_amt_dict()
+    for el in el_amt_dict:
 
-    symbols = sorted(el_amt_dict.keys(), key=lambda sym: get_el_sp(sym).X)
+        if callable(getattr(Element(el), prop)) is None:
+            raise ValueError('Invalid pymatgen Element attribute(property)')
 
-    for el_sym in symbols:
-
-        element = Element(el_sym)
-        property_value = None
-        property_units = None
-
-        try:
-            p = getattr(element, property_name)
-        except AttributeError:
-            print("{} attribute missing".format(property_name))
-            raise
-
-        if p is not None:
-            if property_name in ['ionic_radii']:
-                if oxidation_states:
-                    property_value = element.ionic_radii[oxidation_states[el_sym]]
-                    property_units = Unit("ang")
-                else:
-                    raise ValueError("oxidation state not given for {}; It does not yield a unique "
-                                     "number per Element".format(property_name))
-            else:
-                property_value = float(p)
+        if getattr(Element(el), prop) is not None:
 
             # units are None for these pymatgen descriptors
             # todo: there seem to be a lot more unitless descriptors which are not listed here... -Alex D
@@ -106,13 +342,13 @@ def get_pymatgen_descriptor(composition, property_name):
                 eldata_tup(element=el, propname=prop, propvalue=float(getattr(Element(el), prop)), propunit=units,
                            amt=el_amt_dict[el]))
 
-        # Make a named tuple out of all the available information
-        eldata_tup_lst.append(eldata_tup(element=el_sym, propname=property_name, propvalue=property_value,
-                                         propunit=property_units, amt=el_amt_dict[el_sym]))
+            # Add descriptor values, one for each atom in the compound
+            for i in range(int(el_amt_dict[el])):
+                eldata.append(float(getattr(Element(el), prop)))
 
-        # Add descriptor values, one for each atom in the compound
-        for i in range(int(el_amt_dict[el_sym])):
-            eldata.append(property_value)
+        else:
+            eldata_tup_lst.append(eldata_tup(element=el, propname=prop, propvalue=None, propunit=None,
+                                             amt=el_amt_dict[el]))
 
     return eldata
 
@@ -129,12 +365,10 @@ def get_magpie_descriptor(comp, descriptor_name):
     Returns: (list) of descriptor values for each atom in the composition
 
     """
-    data_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), "data",
-                            'magpie_elementdata')
+    data_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), "data", 'magpie_elementdata')
     magpiedata = []
     magpiedata_tup_lst = []
-    magpiedata_tup = collections.namedtuple('magpiedata_tup',
-                                            'element propname propvalue propunit amt')
+    magpiedata_tup = collections.namedtuple('magpiedata_tup', 'element propname propvalue propunit amt')
     available_props = []
 
     # Make a list of available properties
@@ -144,8 +378,7 @@ def get_magpie_descriptor(comp, descriptor_name):
 
     if descriptor_name not in available_props:
         raise ValueError(
-            "This descriptor is not available from the Magpie repository. Choose from {}".format(
-                available_props))
+            "This descriptor is not available from the Magpie repository. Choose from {}".format(available_props))
 
     # Get units from Magpie README file
     el_amt = Composition(comp).get_el_amt_dict()
@@ -176,61 +409,6 @@ def get_magpie_descriptor(comp, descriptor_name):
             for i in range(int(el_amt[el])):
                 magpiedata.append(propvalue)
 
-    return magpiedata
-
-def get_magpie_dict(comp, descriptor_name):
-    """
-    Gets magpie data for an element. 
-    First checks if magpie properties are already loaded, if not, stores magpie data in a dictionary
-
-    Args: 
-        elem (string): Atomic symbol of element
-        descriptor_name (string): Name of descriptor
-
-    Returns:
-        attr_dict
-    """
-
-    if descriptor_name not in magpie_props:
-
-        data_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), "data", 'magpie_elementdata')
-        available_props = []
-
-        # Make a list of available properties
-        for datafile in os.listdir(data_dir):
-            available_props.append(datafile.replace('.table', ''))
-
-        if descriptor_name not in available_props:
-            raise ValueError(
-                "This descriptor is not available from the Magpie repository. Choose from {}".format(available_props))
-
-    #if descriptor_name not in magpie_props:
-        # Extract from data file
-        prop_value = []
-        with open(os.path.join(data_dir, '{}.table'.format(descriptor_name)), 'r') as descp_file:
-            lines = descp_file.readlines()
-            
-            for atomic_no in range(1, 104): #This is as high as pymatgen goes
-                try:
-                    if descriptor_name in ["OxidationStates"]:
-                        prop_value.append([float(i) for i in lines[atomic_no - 1].split()])
-                    else: 
-                        prop_value.append(float(lines[atomic_no - 1]))
-                except:
-                    prop_value.append(float("NaN"))
-        
-        attr_dict = dict(zip(atomic_syms, prop_value))    
-        magpie_props[descriptor_name] = attr_dict #Add dictionary to magpie_props
-
-    ##Get data for given element/compound
-    el_amt = Composition(comp).get_el_amt_dict()
-    elements = list(el_amt.keys())
-    
-    magpiedata = []
-    for el in elements:
-        for i in range(int(el_amt[el])):
-            magpiedata.append(magpie_props[descriptor_name][el])
- 
     return magpiedata
 
 def get_cohesive_energy(comp):
@@ -307,188 +485,6 @@ def get_holder_mean(data_lst, power):
             total += value ** power
         return (total / len(data_lst)) ** (1 / float(power))
 
-###Stoichiometric attributes from Ward npj paper
-def get_stoich_attributes(comp, p):
-    """
-    Get stoichiometric attributes
-
-    Args:
-        comp (string): Chemical formula
-        p (int)
-    
-    Returns: 
-        p_norm (float): Lp norm-based stoichiometric attribute
-    """
-
-    el_amt = Composition(comp).get_el_amt_dict()
-    
-    p_norm = 0
-    n_atoms = sum(el_amt.values())
-
-    if p < 0:
-        raise ValueError("p-norm not defined for p < 0")
-
-    if p == 0:
-        p_norm = len(el_amt.values())
-    else:
-        for i in el_amt:
-            p_norm += (el_amt[i]/n_atoms)**p
-        p_norm = p_norm**(1.0/p)
-
-    return p_norm
-
-###Elemental properties from Ward npj paper
-def get_elem_property_attributes(comp):
-    """
-    Get elemental property attributes
-
-    Args:
-
-        comp (string): Chemical formula
-    
-    Returns:
-        all_attributes: min, max, range, mean, average deviation, and mode of 22 descriptors
-    """
-
-    comp_obj = Composition(comp)
-    el_amt = comp_obj.get_el_amt_dict()
-    elements = list(el_amt.keys())
-
-    #atom_frac = []
-    #for elem in elements:
-        #atom_frac.append(comp_obj.get_atomic_fraction(elem))
-
-    req_data = ["Number", "MendeleevNumber", "AtomicWeight","MeltingT","Column","Row","CovalentRadius","Electronegativity",
-        "NsValence","NpValence","NdValence","NfValence","NValance","NsUnfilled","NpUnfilled","NdUnfilled","NfUnfilled","NUnfilled",
-        "GSvolume_pa","GSbandgap","GSmagmom","SpaceGroupNumber"]    
-    
-    all_attributes = []
-
-    for attr in req_data:
-        
-        elem_data = get_magpie_dict(comp, attr)
-
-        #for elem in elements:
-        #    elem_data.append(get_magpie_dict(elem, attr)[0])
-
-        desc_stats = []
-        desc_stats.append(min(elem_data))
-        desc_stats.append(max(elem_data))
-        desc_stats.append(max(elem_data) - min(elem_data))
-        
-        prop_mean = np.mean(elem_data)
-        desc_stats.append(prop_mean)
-        desc_stats.append(np.mean(np.abs(np.subtract(elem_data, prop_mean))))
-        desc_stats.append(max(set(elem_data), key=elem_data.count))
-        all_attributes.append(desc_stats)
-
-    return all_attributes
-
-def get_valence_orbital_attributes(comp):
-    """Weighted fraction of valence electrons in each orbital
-
-       Args: 
-            comp (string): Chemical formula
-
-       Returns: 
-            Fs, Fp, Fd, Ff (float): Fraction of valence electrons in s, p, d, and f orbitals
-    """    
-
-    avg_total_valence = np.mean(get_magpie_dict(comp, "NValance"))
-    avg_s = np.mean(get_magpie_dict(comp, "NsValence"))
-    avg_p = np.mean(get_magpie_dict(comp, "NpValence"))
-    avg_d = np.mean(get_magpie_dict(comp, "NdValence"))
-    avg_f = np.mean(get_magpie_dict(comp, "NfValence"))
-
-    Fs = avg_s/avg_total_valence
-    Fp = avg_p/avg_total_valence
-    Fd = avg_d/avg_total_valence
-    Ff = avg_f/avg_total_valence
-
-    return Fs, Fp, Fd, Ff
-
-def get_ionic_attributes(comp):
-    """
-    Ionic character attributes
-
-    Args:
-        comp (string): Chemical formula
-
-    Returns:
-        cpd_possible (bool): Indicates if a neutral ionic compound is possible
-        max_ionic_char (float): Maximum ionic character between two atoms
-        avg_ionic_char (float): Average ionic character
-    """
-    comp_obj = Composition(comp)
-    el_amt = comp_obj.get_el_amt_dict()
-    elements = list(el_amt.keys())
-    values = list(el_amt.values())
- 
-    #Determine if it is possible to form neutral ionic compound
-    ox_states = []
-    for elem in elements:
-        ox_states.append(get_magpie_dict(elem,"OxidationStates")[0])
-    
-    cpd_possible = False
-    ox_sets = itertools.product(*ox_states)
-    for ox in ox_sets:
-        if np.dot(ox, values) == 0:
-            cpd_possible = True
-            break    
-
-    atom_pairs = itertools.combinations(elements, 2)
-
-    ionic_char = []
-    avg_ionic_char = 0
-
-    for pair in atom_pairs:
-        XA = get_magpie_dict(pair[0], "Electronegativity")
-        XB = get_magpie_dict(pair[1], "Electronegativity")
-        ionic_char.append(1.0 - np.exp(-0.25*(XA[0]-XB[0])**2))
-        avg_ionic_char += comp_obj.get_atomic_fraction(pair[0])*comp_obj.get_atomic_fraction(pair[1])*ionic_char[-1]
-    
-    max_ionic_char = np.max(ionic_char)
- 
-    return cpd_possible, max_ionic_char, avg_ionic_char
-
-def get_composition_oxidation_state(formula):
-    """
-    Returns the composition and oxidation states from the given formula.
-    Formula examples: "NaCl", "Na+1Cl-1",   "Fe2+3O3-2" or "Fe2 +3 O3 -2"
-
-    Args:
-        formula (str):
-
-    Returns:
-        pymatgen.core.composition.Composition, dict of oxidation states as strings
-
-    """
-    oxidation_states_dict = {}
-    non_alphabets = re.split('[a-z]+', formula, flags=re.IGNORECASE)
-    if not non_alphabets:
-        return Composition(formula), oxidation_states_dict
-    oxidation_states = []
-    for na in non_alphabets:
-        s = na.strip()
-        if s != "" and ("+" in s or "-" in s):
-            digits = re.split('[+-]+', s)
-            sign_tmp = re.split('\d+', s)
-            sign = [x.strip() for x in sign_tmp if x.strip() != ""]
-            oxidation_states.append("{}{}".format(sign[-1], digits[-1].strip()))
-    if not oxidation_states:
-        return Composition(formula), oxidation_states_dict
-    formula_plain = []
-    before, after = tuple(formula.split(oxidation_states[0], 1))
-    formula_plain.append(before)
-    for oxs in oxidation_states[1:]:
-        before, after = tuple(after.split(oxs, 1))
-        formula_plain.append(before)
-    for i, g in enumerate(formula_plain):
-        el = re.split("\d", g.strip())[0]
-        oxidation_states_dict[str(Element(el))] = int(oxidation_states[i])
-    return Composition("".join(formula_plain)), oxidation_states_dict
-
-
 if __name__ == '__main__':
     descriptors = ['atomic_mass', 'X', 'Z', 'thermal_conductivity', 'melting_point',
                    'coefficient_of_linear_thermal_expansion']
@@ -498,14 +494,15 @@ if __name__ == '__main__':
     print(get_magpie_descriptor('LiFePO4', 'AtomicVolume'))
     print(get_magpie_descriptor('LiFePO4', 'Density'))
     print(get_holder_mean([1, 2, 3, 4], 0))
-    
-    ####TESTING WARD NPJ DESCRIPTORS
+   
+    training_set = pd.DataFrame({"composition":["Fe2O3"]})
     print "WARD NPJ ATTRIBUTES"
     print "Stoichiometric attributes"
-    print get_stoich_attributes("Fe2O3", 3)
+    p_list = [0,2,3,5,7,9]
+    print StoichAttributes().featurize_all(training_set)
     print "Elemental property attributes"
-    print get_elem_property_attributes("Fe2O3")
+    print ElemPropertyAttributes().featurize_all(training_set)
     print "Valence Orbital Attributes"
-    print get_valence_orbital_attributes("Fe2O3")
+    print ValenceOrbitalAttributes().featurize_all(training_set)
     print "Ionic attributes"
-    print get_ionic_attributes("Fe2O3")
+    print IonicAttributes().featurize_all(training_set)
